requirements.txt
dj-database-url==0.5.0
Django==3.1.6
gunicorn==20.0.4
psycopg2-binary==2.7.7
whitenoise==5.1.0
django-import-export==2.3.0
diff-match-patch==20200713
tablib==2.0.0
openpyxl==3.0.4
django-tempus-dominus==5.1.2.13
django-filter==2.3.0
